wtp/webrtc/RTCPeerConnection-constructor.html

<!doctype html>
<meta charset="utf-8">
<title>RTCPeerConnection constructor</title>
<script src="/resources/webrtc-ie-shim.js"></script><script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
'use strict';

test(function () {
  assert_equals(RTCPeerConnection.length, 0);
}, 'RTCPeerConnection.length');

// These are used for string and number dictionary members to see if they are
// being accessed at all.
var toStringThrows = { toString: function toString() {
    throw new Error();
  } };
//var toNumberThrows = Symbol();

// Test the first argument of the constructor. The key is the argument itself,
// and the value is the first argument for assert_throws, or false if no
// exception should be thrown.
var testArgs = {
  // No argument or equivalent.
  '': false,
  'null': false,
  'undefined': false,
  '{}': false,

  // peerIdentity
  '{ peerIdentity: toStringThrows }': new Error(),

  // certificates
  '{ certificates: null }': new TypeError(),
  '{ certificates: undefined }': false,
  '{ certificates: [] }': false,
  '{ certificates: [null] }': new TypeError(),
  '{ certificates: [undefined] }': new TypeError(),

  // iceCandidatePoolSize
  //'{ iceCandidatePoolSize: toNumberThrows }': new TypeError()
};

var _loop = function _loop(arg) {
  var expr = 'new RTCPeerConnection(' + arg + ')';
  test(function () {
    var throws = testArgs[arg];
    if (throws) {
      assert_throws(throws, function () {
        eval(expr);
      });
    } else {
      eval(expr);
    }
  }, expr);
};

for (var arg in testArgs) {
  _loop(arg);
}

// The initial values of attributes of RTCPeerConnection.
var initialState = {
  'localDescription': null,
  'currentLocalDescription': null,
  'pendingLocalDescription': null,
  'remoteDescription': null,
  'currentRemoteDescription': null,
  'pendingRemoteDescription': null,
  'signalingState': 'stable',
  'iceGatheringState': 'new',
  'iceConnectionState': 'new',
  'connectionState': 'new',
  'canTrickleIceCandidates': null
  // TODO: defaultIceServers
};

var _loop2 = function _loop2(attr) {
  test(function () {
    // Use one RTCPeerConnection instance for all initial value tests.
    if (!window.pc) {
      window.pc = new RTCPeerConnection();
    }
    assert_equals(window.pc[attr], initialState[attr]);
  }, attr + ' initial value');
};

for (var attr in initialState) {
  _loop2(attr);
}
</script>
